Continue to rename classes
