+++ src/Core/Domain/Security/CommandHandler/ClearOutdatedCustomerSessionHandlerInterface.php
@@ -24,36 +24,19 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-declare(strict_types=1);
+namespace PrestaShop\PrestaShop\Core\Domain\Security\CommandHandler;
 
-namespace PrestaShop\PrestaShop\Adapter\Security\CommandHandler;
-
-use PrestaShop\PrestaShop\Adapter\Session\Repository\CustomerSessionRepository;
 use PrestaShop\PrestaShop\Core\Domain\Security\Command\ClearOutdatedCustomerSessionCommand;
-use PrestaShop\PrestaShop\Core\Domain\Security\CommandHandler\ClearCustomerSessionHandlerInterface;
 
 /**
- * Class ClearCustomerSessionHandler
- *
- * @internal
+ * Interface ClearOutdatedCustomerSessionHandlerInterface defines session deletion handler.
  */
-final class ClearCustomerSessionHandler implements ClearCustomerSessionHandlerInterface
+interface ClearOutdatedCustomerSessionHandlerInterface
 {
     /**
-     * @var CustomerSessionRepository
+     * Delete session.
+     *
+     * @param ClearOutdatedCustomerSessionCommand $command
      */
-    private $repository;
-
-    public function __construct(CustomerSessionRepository $repository)
-    {
-        $this->repository = $repository;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function handle(ClearOutdatedCustomerSessionCommand $command): void
-    {
-        $this->repository->clearOutdatedSessions();
-    }
+    public function handle(ClearOutdatedCustomerSessionCommand $command): void;
 }
