+++ src/PrestaShopBundle/Resources/config/services/adapter/security.yml
@@ -11,12 +11,12 @@ services:
         command: 'PrestaShop\PrestaShop\Core\Domain\Security\Command\DeleteCustomerSessionCommand'
 
   prestashop.adapter.security.command_handler.clear_customer_session_handler:
-    class: 'PrestaShop\PrestaShop\Adapter\Security\CommandHandler\ClearCustomerSessionHandler'
+    class: 'PrestaShop\PrestaShop\Adapter\Security\CommandHandler\ClearOutdatedCustomerSessionHandler'
     arguments:
       - '@prestashop.adapter.security.repository.customer_session_repository'
     tags:
       - name: 'tactician.handler'
-        command: 'PrestaShop\PrestaShop\Core\Domain\Security\Command\ClearCustomerSessionCommand'
+        command: 'PrestaShop\PrestaShop\Core\Domain\Security\Command\ClearOutdatedCustomerSessionCommand'
 
   prestashop.adapter.security.command_handler.delete_employee_session_handler:
     class: 'PrestaShop\PrestaShop\Adapter\Security\CommandHandler\DeleteEmployeeSessionHandler'
@@ -27,12 +27,12 @@ services:
         command: 'PrestaShop\PrestaShop\Core\Domain\Security\Command\DeleteEmployeeSessionCommand'
 
   prestashop.adapter.security.command_handler.clear_employee_session_handler:
-    class: 'PrestaShop\PrestaShop\Adapter\Security\CommandHandler\ClearEmployeeSessionHandler'
+    class: 'PrestaShop\PrestaShop\Adapter\Security\CommandHandler\ClearOutdatedEmployeeSessionHandler'
     arguments:
       - '@prestashop.adapter.security.repository.employee_session_repository'
     tags:
       - name: 'tactician.handler'
-        command: 'PrestaShop\PrestaShop\Core\Domain\Security\Command\ClearEmployeeSessionCommand'
+        command: 'PrestaShop\PrestaShop\Core\Domain\Security\Command\ClearOutdatedEmployeeSessionCommand'
 
   prestashop.adapter.security.command_handler.bulk_delete_employees_sessions_handler:
     class: 'PrestaShop\PrestaShop\Adapter\Security\CommandHandler\BulkDeleteEmployeeSessionsHandler'
